Fixing the FAB animation, and other minor improvements
Turns out the center point of the FAB animation was slightly off due to
it needing to be scaled by the screen density. Also slowed down the
animation a bit, and fixed up a few other things in the UI (like hiding
the toolbar items when the drawer is open).