Made "search zips" per-folder rather than global
“Search zips” is now applied to each folder individually, rather than
globally. This required removing the text prompt to select a file OR a
folder, but they should be able to figure that out on their own.

Importing has also been optimized considerably in certain cases,
although it now batches ZIP files together after importing the
non-zipped files. That’s only because I needed a way to apply the
per-folder “search zips” preference, and that’s the method I chose.