Fixed keyboard/D-pad navigation
The app appears to be completely usable using the keyboard or D-pad
again. Was a simple matter of setting views to focusable and assigning
a standard drawable as their backgrounds. Although I also did have to
fix a view reuse bug which caused the navigation drawer’s views to all
turn gray eventually.