Now uses RecyclerView and GridLayoutManager to save memory
Instead of loading all of the ROM thumbnails in at once, RecyclerView
handles loading only the thumbnails that will be visible on the screen.
As the user scrolls, thumbnails are loaded in and out. Also switched to
GridLayoutManager and implemented thumbnails that automatically resize
to fit the available space. Still need to implement proper item spacing.