Bug fix with closing the sidebar in the middle of it opening
If the user closed the game sidebar when it was in the middle of
animating to open, the next time they opened the sidebar it would
contain the game information instead of the navigation menu. This is
because onDrawerClosed is not called. The trick is to detect this
intercepted state and wait until the drawer state changes to idle.