Prevent accidentally showing the game sidebar
Was thoroughly testing the new game sidebar to make sure it didn’t
accidentally trigger during gameplay, and… it did. Was a lot more
common with Kirby 64 due to the D-pad, but it happened with the analog
stick sometimes too. This commit prevents accidentally opening the
drawer, although I’m not 100% happy with it yet. I may change it to
allow edge swipes over normal buttons since it’s highly unusual to
swipe horizontally along them, and because placing a bunch of buttons
along the left edge makes it impossible to show the sidebar.